// frontend/src/api/models/librarySourceDto.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface LibrarySourceDto {
  sourceId?: string;
  trackId?: string;
  title?: string;
  artist?: string;
  album?: string;
  provider?: string;
  addedAt?: string;
  linkedToCatalog?: boolean;
}